+++ editors/code/src/extension.ts
@@ -182,12 +182,12 @@ export async function activate(context: vscode.ExtensionContext) {
       { scheme: "file", language: "lua" },
       { scheme: "file", language: "luau" },
     ],
-    synchronize: {
-      fileEvents: [
-        vscode.workspace.createFileSystemWatcher("**/.luaurc"),
-        vscode.workspace.createFileSystemWatcher("**/sourcemap.json"),
-      ],
-    },
+    // synchronize: {
+    //   fileEvents: [
+    //     vscode.workspace.createFileSystemWatcher("**/.luaurc"),
+    //     vscode.workspace.createFileSystemWatcher("**/sourcemap.json"),
+    //   ],
+    // },
     diagnosticCollectionName: "luau",
   };